Fixed SOLID stuff
