+++ kerem-net/src/Features/Post/Comment/Comment.ts
@@ -1,0 +1,6 @@
+export interface Comment {
+  id: string;
+  author: string;
+  content: string;
+  date: Date;
+}
