+++ kerem-net/src/Features/Post/Post/Post.ts
@@ -1,0 +1,10 @@
+import { Comment } from "../Comment/Comment";
+
+export interface Post {
+  id: string;
+  author: string;
+  content: string;
+  date: Date;
+  likes: number;
+  comments: Comment[];
+}
